+++ core/src/ldp-container/LdpContainer.ts
@@ -1,6 +1,7 @@
 import { sym } from "rdflib";
 import { labelFromUri, Thing } from "../thing";
 import { Store } from "../Store";
+import { debounceTime, filter, map, merge, Observable, startWith } from "rxjs";
 
 export interface ContainerContent {
   uri: string;
@@ -15,6 +16,11 @@ export class LdpContainer extends Thing {
     super(uri, store, editable);
   }
 
+  /**
+   * Resources that the LDP Container contains
+   *
+   * @returns Array of objects with uri and name
+   */
   contains(): ContainerContent[] {
     const contains = this.store.statementsMatching(
       sym(this.uri),
@@ -26,5 +32,23 @@ export class LdpContainer extends Thing {
       uri: content.object.value,
       name: labelFromUri(content.object.value),
     }));
+  }
+
+  /**
+   * Observe changes to the resources that the LDP Container contains
+   *
+   * @returns RxJS Observable that pushes a new contains() array when it changes
+   */
+  observeContains(): Observable<ContainerContent[]> {
+    return merge(this.store.additions$, this.store.removals$).pipe(
+      filter(
+        (quad) =>
+          quad.graph.value == this.uri &&
+          quad.predicate.value == "http://www.w3.org/ns/ldp#contains",
+      ),
+      debounceTime(250),
+      map(() => this.contains()),
+      startWith(this.contains()),
+    );
   }
 }